eager data loading ready
